+++ client/src/index.css
@@ -6,149 +6,3 @@
 div {
   font-family: public-pixel;
 }
-
-body {
-  background-image: url(grass.png);
-  background-size: 70%;
-}
-
-.info-box {
-  position: fixed;
-  top: 0;
-  right: 0;
-  height: 160px;
-  width: 245px;
-  margin: 10px;
-}
-
-.time {
-  position: absolute;
-  top: 30px;
-  right: 30px;
-}
-
-@keyframes spin {
-  0% {rotate: 0deg;}
-  100% {rotate: 360deg;}
-}
-
-.icon {
-  margin-right: 30px;
-  animation-name: spin;
-  animation-duration: 20s;
-  animation-delay: 0s;
-  animation-timing-function: linear;
-  animation-iteration-count: infinite;
-
-}
-
-.pet-container {
-  position: relative;
-  display: inline-block;
-  transition: top 100s, left 100s; /* Transition effect for smooth movement */
-}
-
-.details-popup {
-  position: absolute;
-  top: 0;
-  left: 100%;
-  width: 200px; /* Adjust the width as needed */
-  padding: 10px;
-  background-color: #f0f0f0;
-  border: 1px solid #ccc;
-}
-
-@keyframes petAnimation {
-  0% {
-    transform: scale(1);
-  }
-  50% {
-    transform: scale(1.1);
-  }
-  100% {
-    transform: scale(1);
-  }
-}
-
-#pet {
-  animation: petAnimation 2s linear infinite;
-}
-
-.pet-container.running {
-  animation: petMove 100s linear infinite;
-}
-
-@keyframes petMove {
-  0% {
-    transform: translate(0, 0);
-  }
-  50% {
-    transform: translate(50px, 50px);
-  }
-  100% {
-    transform: translate(0, 0);
-  }
-}
-
-.clock {
-  position: fixed;
-  top: 10px;
-  right: 10px;
-  display: flex;
-  align-items: center;
-  color: #ffffff;
-  font-family: Arial, sans-serif;
-  font-size: 16px;
-}
-
-.time {
-  font-weight: bold;
-}
-
-.pet-count {
-  margin-top: 50px;
-}
-
-.grave {
-  position: relative;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  overflow: hidden;
-}
-
-.grave-info {
-  position: absolute;
-  top: 0;
-  left: 100%;
-  width: 200px;
-  height: 100%;
-  padding: 10px;
-  background-color: rgba(0, 0, 0, 0.8);
-  color: #ffffff;
-  display: none;
-}
-
-.grave:hover .grave-info {
-  display: flex;
-  flex-direction: column;
-}
-
-.pet-image {
-  width: 100px;
-  height: 100px;
-  object-fit: cover;
-}
-
-.pet-name {
-  font-weight: bold;
-  margin-top: 10px;
-}
-
-.death-date {
-  margin-top: 5px;
-}
-
-.fun-fact {
-  margin-top: 5px;
-}
